backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Adaptive RAG Backend with LangGraph and FastAPI"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Web 框架
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    # LangChain 核心
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-text-splitters>=0.3.0",
    # 向量数据库
    "chromadb>=0.4.0",
    # 本地嵌入模型
    "langchain-huggingface>=0.1.0",
    "sentence-transformers>=2.2.0",
    # 搜索工具
    "langchain-community>=0.3.0",
    # 文档加载
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    # 环境变量管理
    "python-dotenv>=1.0.0",
    # HTTP 客户端
    "httpx>=0.25.0",
]
